`pingod.windows`: re factored plug-in. Added `tools_panel`, `ToolsPanelWindow`, this overlays buttons you can add to activate different windows on top of the game, by default these are embedded windows but if you were using a smaller display you wouldn't have them embedded.
added `ConsoleWindow`, this is hack to get the latest log. When you are full screen or maximized window debugging this is better than shifting back to other windows.

NOTE: used to be `_switchWindoEnabled` now `ToolsWindowEnabled`

TODO: added better documentation on how to quickly add a new window and tool button to open it